+++ src/read_write/write_mmcif.jl
@@ -57,6 +57,8 @@ function _supported_write_cif_fields(field_assignment)
     return replace_custom_fields!(_atom_symbol_for_cif_field, field_assignment)
 end
 
+_dot_for_empty(val) = val
+_dot_for_empty(val::AbstractString) = isnothing(findfirst(!isspace, val)) ? "." : val
 function _get_mmcif_field(atom, field_name)
     field_name == :_print_atom && return "ATOM"
     field_name == :_print_dot && return "."
@@ -68,7 +70,7 @@ function _get_mmcif_field(atom, field_name)
         return 3
     end
     if field_name in fieldnames(typeof(atom))
-        return getfield(atom, field_name)
+        return _dot_for_empty(getfield(atom, field_name))
     end
     throw(ArgumentError(":$field_name not found for printing mmcif field in atom"))
 end
@@ -85,7 +87,7 @@ The optional `field_assignment` argument is a dictionary that can be used to ass
 function write_mmcif(
     filename::AbstractString,
     atoms::AbstractVector{<:Atom},
-    selection::String;
+    selection::AbstractString;
     field_assignment::Union{Nothing,Dict{String,Symbol}}=nothing,
 )
     write_mmcif(filename, atoms, parse_query(selection); field_assignment)
@@ -222,4 +224,10 @@ end
     @test isfile(tmpfile)
     ats_cif = read_mmcif(tmpfile)
     @test all(position(at1) ≈ position(at2) for (at1, at2) in zip(cA, ats_cif))
+
+    at = Atom(name="HC", pdb_element="")
+    write_mmcif(tmpfile, [at])
+    ats = read_mmcif(tmpfile)
+    @test element(ats[1]) == "H"
+
 end