+++ CHANGELOG.md
@@ -10,8 +10,16 @@ PDBTools.jl Changelog
 [badge-fix]: https://img.shields.io/badge/Fix-purple.svg
 [badge-info]: https://img.shields.io/badge/Info-gray.svg
 
-Version 3.23.2-DEV
+Version 3.24.2-DEV
 --------------
+
+Version 3.24.1
+--------------
+- ![FIX][badge-fix] Missing data is printed as dot in mmCIF files, avoiding missing columns, and the dot is recognized as a missing `pdb_element` field by the `element` function.
+
+Version 3.24.0
+--------------
+- ![FEATURE][badge-feature] Support for `AbstractString` instead of `String` in several input signatures, notably for selections and file names.
 
 Version 3.23.1
 --------------